refactor: consolidate duplicate ErrorBoundary components
Consolidate duplicate ErrorBoundary implementations from
components/error-boundary.tsx and lib/error-handler.tsx into a
single unified component.

The new ErrorBoundary in components/error-boundary.tsx supports:
- Simple fallback (ReactNode)
- Component fallback with resetError callback
- Optional onReset callback
- showReload prop to switch between "Try Again" and "Reload Page" buttons

The lib/error-handler.tsx now re-exports ErrorBoundary from the
components directory for convenience while keeping its toast helper
functions.
