+++ website/lib/error-handler.tsx
@@ -34,65 +34,6 @@ export const showInfo = (message: string) => {
   toast.info(message);
 };
 
-// Error boundary component
-'use client';
-
-import * as React from 'react';
-import { Button } from '@/components/ui/button';
-
-interface ErrorBoundaryProps {
-  children: React.ReactNode;
-  fallback?: React.ComponentType<{ error: Error; resetError: () => void }>;
-}
-
-interface ErrorBoundaryState {
-  hasError: boolean;
-  error: Error | null;
-}
-
-export class ErrorBoundary extends React.Component<ErrorBoundaryProps, ErrorBoundaryState> {
-  constructor(props: ErrorBoundaryProps) {
-    super(props);
-    this.state = { hasError: false, error: null };
-  }
-
-  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
-    return { hasError: true, error };
-  }
-
-  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
-    console.error('Error caught by boundary:', error, errorInfo);
-    showError(error, 'UI Component Error');
-  }
-
-  resetError = () => {
-    this.setState({ hasError: false, error: null });
-  };
-
-  render() {
-    if (this.state.hasError) {
-      if (this.props.fallback) {
-        const FallbackComponent = this.props.fallback;
-        return <FallbackComponent error={this.state.error!} resetError={this.resetError} />;
-      }
-
-      return (
-        <div className="min-h-screen flex items-center justify-center bg-background">
-          <div className="max-w-md w-full space-y-4 p-6 rounded-lg border border-border bg-card">
-            <div className="text-center">
-              <h2 className="text-lg font-semibold text-foreground mb-2">Something went wrong</h2>
-              <p className="text-muted-foreground mb-4">
-                {this.state.error?.message || 'An unexpected error occurred'}
-              </p>
-              <Button onClick={this.resetError} variant="default">
-                Try Again
-              </Button>
-            </div>
-          </div>
-        </div>
-      );
-    }
-
-    return this.props.children;
-  }
-}
+// Re-export ErrorBoundary from components for convenience
+export { ErrorBoundary as ErrorBoundary } from '@/components/error-boundary';
+export { default as ErrorBoundaryDefault } from '@/components/error-boundary';
